Split review packet smoke fixtures
